Extract shared sim-factory and input-utils for test reuse
- sim-factory.js: createSimFighter, createSimCombat, resetSimFighter
  extracted from rollback-round-events.test.js (~340 lines removed)
- input-utils.js: expandSparseInputs for converting sparse input logs
  to dense frame-indexed arrays
- rollback-round-events.test.js: imports from sim-factory instead of
  inline definitions
